remove unnecessary tests
