+++ components/org.wso2.carbon.identity.scim2.common/src/test/resources/testng.xml
@@ -28,7 +28,6 @@
             <class name="org.wso2.carbon.identity.scim2.common.impl.IdentitySCIMManagerTest"/>
             <class name="org.wso2.carbon.identity.scim2.common.utils.AdminAttributeUtilTest"/>
             <class name="org.wso2.carbon.identity.scim2.common.impl.SCIMUserManagerTest"/>
-            <class name="org.wso2.carbon.identity.scim2.common.internal.SCIMCommonComponentHolderTest"/>
             <class name="org.wso2.carbon.identity.scim2.common.internal.SCIMCommonComponentTest"/>
 
             <class name="org.wso2.carbon.identity.scim2.common.impl.SCIMRoleManagerTest"/>
